requirements.txt
streamlit==1.37.1
requests
openai==1.37.0
httpx==0.27.2
python-dotenv
